Use file for input data instead of STDIN ( non-interactive testing )
test.pl as-is is useless because it executes no readline code when in
automated testing conditions.

Additionally, attempting to use the built-in file redirection
mechanisms is useless, as there's no way to set those parts from

    make test

And 'preput' support causes readline to barf when the input stream
is not a real TTY.

Subsequently, the pre-inserted "exit" also has to be removed.

Bug: https://bugs.gentoo.org/492212